third_party/blink/web_tests/external/wpt/webidl/ecmascript-binding/builtin-function-properties.any.worker-expected.txt

This is a testharness.js-based test.
[FAIL] Constructor property enumeration order of "length", "name", and "prototype"
  assert_array_equals: expected property 2 to be "prototype" but got "arguments" (expected array ["length", "name", "prototype"] got ["length", "name", "arguments"])
Harness: the test ran to completion.
